Improve example documentation
